getNextLowerBlockNumber private
